feat: Add Candidate and RecruitmentStage models, seeders, and migrations
