+++ app/Models/Candidate.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Candidate extends Model
+{
+    //
+}
